Allow pager to consume drag events at bounds
This change returns early from the drag event
when a zoomed image is at its maximum X-axis
bounds. This ensures that the drag event is not
consumed at is available to parent components
like a view pager.

The isMaxX function ensures that the current
drag event is in the same direction as the
current ZoomableState translationX. If the
bounds are within one pixel then the isMaxX
function returns true.